+++ lib/aden.js
@@ -450,6 +450,10 @@ AdenBackend.prototype.shutdown = function shutdown(done) {
           delete require.cache[key];
         });
 
+      if (this.server) {
+        this.server.close();
+      }
+
       done();
     });
   }
