+++ lib/aden.logger.js
@@ -246,7 +246,7 @@ LogParser.prototype.getStreamHandler = function getStreamHandler() {
     }
     lines.forEach((line) => {
       const lineType = getLineEventType(line);
-
+      console.log(line);
       if (lineType) {
         const jsonData = this.parseLine(line, lineType);
 
@@ -283,7 +283,7 @@ LogParser.prototype.getStreamHandler = function getStreamHandler() {
           if (this.listeners('error').length > 0) {
             this.emit('error', err, jsonData);
           }
-          console.log('ci test', err, jsonData);
+
           if (err.message && err.message.match(/Worker error code/i)) {
             this.emit('worker:error', err, jsonData);
           }
